+++ Code/JavaWrappers/csharp_wrapper/CMakeLists.txt
@@ -1,17 +1,6 @@
 project (GraphMolCSharp)
 
 include_directories( ${RDKit_ExternalDir} )
-
-# find the mcs executables on non-windows systems:
-if(NOT MSVC)
-  find_program(GMCS_EXE mcs)
-  if (NOT GMCS_EXE)
-	MESSAGE ("mcs (executable) is not found. Please add it to PATH and rerun cmake.")
-	MESSAGE(FATAL_ERROR "Cannot find required executable mcs")
-  endif (NOT GMCS_EXE)
-endif(NOT MSVC)
-
-
 SET_SOURCE_FILES_PROPERTIES(GraphMolCSharp.i PROPERTIES CPLUSPLUS ON )
 
 # Setup a few variables for environment-specific things
@@ -92,22 +81,32 @@ SWIG_LINK_LIBRARIES(RDKFuncs ${RDKit_Wrapper_Libs}
 INSTALL(TARGETS RDKFuncs
         DESTINATION ${CMAKE_CURRENT_SOURCE_DIR} )
 
+if (RDK_BUILD_DOTNET_CSHARP_TESTS)
 
-if(NOT MSVC)
-   # code adapted from the wrapper code for
-   # GDCM: http://gdcm.svn.sf.net/viewvc/gdcm/trunk/Wrapping/Java/CMakeLists.txt?view=markup
-   ADD_CUSTOM_COMMAND(
-     OUTPUT ${CMAKE_CURRENT_SOURCE_DIR}/RDKit2DotNet.dll
+  find_program(DOTNET_EXE dotnet)
+  if (NOT DOTNET_EXE)
+    MESSAGE("dotnet (executable) is not found.  Pleae add it to PATH and rerun cmake.")
+    MESSAGE(FATAL_ERROR "Cannot find required executable dotnet")
+  endif (NOT DOTNET_EXE)
 
-     COMMAND ${CMAKE_COMMAND} -E make_directory swig_csharp
+  ADD_CUSTOM_COMMAND(
+    OUTPUT RDKitDotNetLib SYMBOLIC
+  
+    COMMAND ${CMAKE_COMMAND} -E copy $<TARGET_FILE:RDKFuncs> ${CMAKE_CURRENT_SOURCE_DIR}
+    COMMAND ${DOTNET_EXE} build RDKitDotNetCore.sln
+    WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
+    DEPENDS RDKFuncs
+  )
+  
+  ADD_CUSTOM_TARGET(RDKitDotNet ALL
+     DEPENDS RDKFuncs RDKitDotNetLib
+     COMMENT "Building RDKit DotNet project"
+  )
+  
+  ADD_TEST(
+    NAME CSharpTests
+    COMMAND ${DOTNET_EXE} test RDKitDotNetCore.sln "-l:console$<SEMICOLON>verbosity=normal" 
+    WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
+  )
 
-     ## 1. run this custom command only after swig has been run.
-     COMMAND ${GMCS_EXE} -platform:${PLATFORM} -out:RDKit2DotNet.dll -t:library "swig_csharp/*.cs"
-       WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
-     DEPENDS "${swig_generated_file_fullname}"
-   )
-   ADD_CUSTOM_TARGET(RDKFuncsDLL ALL
-     DEPENDS RDKFuncs ${CMAKE_CURRENT_SOURCE_DIR}/RDKit2DotNet.dll
-     COMMENT "building mono dll"
-   )
-endif(NOT MSVC)
+endif()
