+++ Code/JavaWrappers/csharp_wrapper/RdkitTests/MolEnumeratorTest.cs
