+++ Code/JavaWrappers/csharp_wrapper/RdkitTests/Usings.cs
@@ -1,0 +1,1 @@
+global using Xunit;
