+++ Code/JavaWrappers/csharp_wrapper/RdkitTests/xunit.runner.json
@@ -1,0 +1,4 @@
+{
+  "parallelizeTestCollections": false,
+  "maxParallelThreads": -1
+}
